Revises mismatched type guard.
The intention of this guard is to fail consistently when the type of
a value in the index doesn't match the type stored on the instance;
e.g., in some cases the instance might hold a 'datetime.datetime'
object while the index might hold a 'datetime.date' object.

This guard is extended to ensure this check will always fail when
the type of either the index or instance value is 'NoneType'.